backend/requirements.txt
Flask==3.0.3
flask
psycopg2-binary